move apilib outside of consumerServer
